pyproject.toml
[tool.poetry]
name = "d9.11_homework"
version = "0.1.0"
description = ""
authors = ["azat715 <azat715@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2"
djangorestframework = "^3.12.4"
drf-yasg = "^1.20.0"
gunicorn = "^20.1.0"
whitenoise = "^5.2.0"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pylint = "^2.7.4"
pytest-django = "^4.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
